# common/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "voice-engine",
    "description": "\nStandalone voice service. A frontend, script, or this page only needs the\nserver address (`http://host:port`). Nothing else is shared.\n\n## Try it here, end to end\n\n1. **POST /v1/jobs** \u2014 pick the **design** example, Execute. Copy `id` from the response.\n2. **GET /v1/jobs/{job_id}** \u2014 Execute until `status` is `succeeded`. Each output has a `url`.\n3. **GET /v1/jobs/{job_id}/files/{name}** \u2014 pass that file name (`candidate-01.wav`). The response is a downloadable WAV.\n4. **POST /v1/jobs** \u2014 **lock** example. Set `from_job` to the design id and `candidate` to the one you liked.\n5. **POST /v1/jobs** \u2014 **speak** example, with the same `voice_id`. Download `audio.wav` the same way.\n\nIf the server was started with a token, click **Authorize** and paste it first.\n`GET /health` never needs a token; use it to confirm the address is right.\n\nThe machine-readable spec is `/openapi.json`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit a design, lock, or speak job",
        "operationId": "submit_job_v1_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DesignSpec"
                  },
                  {
                    "$ref": "#/components/schemas/LockSpec"
                  },
                  {
                    "$ref": "#/components/schemas/SpeakSpec"
                  },
                  {
                    "$ref": "#/components/schemas/RenderSpec"
                  }
                ],
                "title": "Spec"
              },
              "examples": {
                "design": {
                  "summary": "1. Design candidate voices",
                  "description": "Writes candidate-01.wav onward. Poll the job, then download a candidate.",
                  "value": {
                    "type": "design",
                    "instruct": "An adult male narrator in his mid-thirties with a warm, low-mid voice. Calm and curious.",
                    "text": "Give it a second. Ordinary things get interesting when they refuse to behave.",
                    "language": "English",
                    "candidates": 2,
                    "seed_start": 1000,
                    "params": {
                      "temperature": 0.9
                    }
                  }
                },
                "lock": {
                  "summary": "2. Lock one candidate",
                  "description": "Replace from_job with the design job id. candidate is 1 for candidate-01.wav.",
                  "value": {
                    "type": "lock",
                    "voice_id": "narrator-male-v1",
                    "from_job": "job_REPLACE_ME",
                    "candidate": 1,
                    "name": "Narrator (male)"
                  }
                },
                "speak": {
                  "summary": "3. Speak new text with the locked voice",
                  "description": "Download audio.wav from the finished job.",
                  "value": {
                    "type": "speak",
                    "voice_id": "narrator-male-v1",
                    "text": "What actually happens inside a combustion engine? Let's slow it down and look.",
                    "language": "English",
                    "seed": 1,
                    "params": {
                      "temperature": 0.8
                    }
                  }
                },
                "render": {
                  "summary": "4. Long script in one voice",
                  "description": "Up to about ten minutes. Tags such as [laughs] and [whispers] are kept on their beat. Download audio.wav.",
                  "value": {
                    "type": "render",
                    "voice_id": "narrator-male-v1",
                    "style": "narration",
                    "text": "[calm] The room went quiet.\n\n[laughs] Then someone at the back started it.\n[whispers] Don't look now.",
                    "language": "English",
                    "seed": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs, newest first",
        "operationId": "list_jobs_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Read a job, including download URLs",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel a job that is still queued",
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/record": {
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete a job and its audio",
        "operationId": "delete_job_v1_jobs__job_id__record_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/files/{name}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Download a job WAV",
        "operationId": "download_job_file_v1_jobs__job_id__files__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WAV file. The browser and Swagger UI offer it as a download.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/designs/{design_id}": {
      "get": {
        "tags": [
          "Voice design"
        ],
        "summary": "Read a design and its candidates",
        "operationId": "get_design_v1_designs__design_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Design Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/renders": {
      "post": {
        "tags": [
          "Speech"
        ],
        "summary": "Render a long script in one voice",
        "operationId": "create_render_v1_renders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Set this only when the server was started with a token.",
        "scheme": "bearer"
      }
    }
  },
...
}